sdk: 7.x spec cleanup
