+++ packages/sdk/test/data/index.ts
@@ -1,0 +1,4 @@
+export * from './stablePools';
+export * from './xykPool';
+export * from './xykPools';
+export * from './xykPoolsDev';
